Make vertical rhythm more robust
Each unit should be as self-contained as possible, and not disrupt the
vertical rhythm for the following elements. There are still a couple of
cases where one element can affect another adversely, but these are
mostly cases with something above the h1. The only one of these I
really care for is `.highlight + hr`, for which a fix is included.