Update Jekyll config to match GitHub Pages
As detailed in [GitHub's blog
post](https://github.com/blog/2289-publishing-with-github-pages-now-as-easy-as-1-2-3),
GitHub Pages now supports the following plugins:

- jekyll-default-layout
- jekyll-optional-front-matter
- jekyll-readme-index
- jekyll-titles-from-headings

All of these are relevant to my interests, and will allow me to
simplify a bunch of my site's pages. To use them locally, though, I
need to disable safe mode, so this commit also does that.