Make archive easier to read
Liquid 3 doesn't support whitespace control, but it *does* allow
arbitrary whitespace within tags, so we can use that to make the control
flow here easier to read.

I've also moved the heading rendering code to the start of the loop, so
it's hopefully a little clearer what the structure of this page will be
when rendered and removes the initial special case.