Move supporting Mac-keyboard-on-Windows files here
There are some supporting files which, while possibly out of date, may
be of interest to someone stumbling on this post, so we should have
them here, and link directly to them.

I'm not *wild* about the path, but this is the easiest way to ensure no
links are broken that might be elsewhere on the web.